Debug decoding for HttpError (plain text)
