+++ ic-agent/src/lib.rs
@@ -102,7 +102,10 @@ pub mod export;
 pub mod identity;
 pub mod request_id;
 
-pub use agent::{agent_error::AgentError, nonce::NonceFactory, Agent, PasswordManager};
+pub use agent::{
+    agent_error::AgentError, agent_error::HttpErrorPayload, nonce::NonceFactory, Agent,
+    PasswordManager,
+};
 pub use identity::{Identity, Signature};
 pub use request_id::{to_request_id, RequestId, RequestIdError};
